+++ test_fixes/kalender_members.dart
@@ -2,6 +2,7 @@
 
 import 'package:flutter/widgets.dart';
 import 'package:kalender/kalender.dart';
+import 'package:kalender/kalender_extensions.dart';
 
 Widget build(BuildContext context, KalenderController c, EventsController e) =>
     KalenderView(calendarController: c, eventsController: e);
@@ -9,3 +10,4 @@ Widget build(BuildContext context, KalenderController c, EventsController e) =>
 KalenderController field(KalenderView view) => view.calendarController;
 KalenderController of(BuildContext context) => KalenderScope.calendarControllerOf(context);
 KalenderController? maybe(BuildContext context) => KalenderScope.maybeCalendarControllerOf(context);
+InternalDateTimeRange? visible(ViewController c) => c.visibleDateTimeRange.value;